I found this very difficult
I sat working on this for a while and scouring the internet. I eventually came across a very similar assignment online that gave me most of the answers with moderate changes. I did my best to understand how the code provided worked though I am still a bit confused on one or two of the questions.